refactor: 📦 Move onboarding interface under api folder
