+++ apps/wizarr-frontend/src/modules/settings/pages/Onboarding.vue
@@ -39,7 +39,7 @@ import OnboardingSection from "../components/Onboarding/OnboardingSection.vue";
 import EditOnboarding from "../components/Modals/EditOnboarding.vue";
 
 import type { Themes } from "md-editor-v3";
-import type { OnboardingPage } from "@/types/OnboardingPage";
+import type { OnboardingPage } from "@/types/api/onboarding/OnboardingPage";
 
 export default defineComponent({
     name: "Onboarding",
