+++ apps/wizarr-frontend/src/types/api/onboarding/onboardingPage.ts
